# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CEX Public & Admin API
  version: 1.0.0
  summary: REST/JSON API for the centralized crypto exchange (frozen DB v1.0).
  description: |
    Contract-first API for the exchange. Generated clients (Flutter, admin SPA)
    are the consumers; **this spec is the boundary**.

    ## Conventions
    - **Transport:** REST/JSON over HTTPS only. UTC ISO-8601 timestamps.
    - **Envelope (success):** `{ "success": true, "data": <T>, "meta"?: {...} }`
    - **Envelope (error):** `{ "success": false, "error": { "code", "message", "details"? } }`
      Clients **switch on `error.code`** (stable enum), never on prose or HTTP status alone.
    - **Money:** all monetary values are **decimal strings** (e.g. `"100.50"`), never
      JSON numbers — mirrors the `NUMERIC` ledger and avoids float drift. Pattern
      `^-?\d+(\.\d+)?$`. Scale is asset-defined (INR=2, USDT=6, …).
    - **IDs:** UUID v4 strings unless noted (`trades.seq` and log ids are int64 strings).
    - **Pagination:** cursor-based — `?cursor=&limit=` → `data.items[]` + `data.nextCursor`
      (null at end). No offset paging (unstable for append-only history).
    - **Versioning:** URL-major (`/api/v1`, `/admin/v1`). See `VERSIONING` in API_DESIGN.md.

    ## Authentication
    - **UserBearer** — short-lived access JWT (10–15 min) in `Authorization: Bearer`.
      Rotate via `POST /auth/refresh` (rotating refresh token, reuse-detected).
...

    ## Idempotency
...
    it carry `x-idempotent: required`.

...
servers:
...
security:
  - UserBearer: []

# ============================================================================
# PATHS — PUBLIC API (/api/v1)
# ============================================================================
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      summary: Rotate tokens using a refresh token.
      security: []
      x-rate-limit: { policy: auth, limit: 30, window: 1m }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } }
      responses:
        '200':
          description: New token pair (old refresh invalidated).
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenPairEnvelope' } } }
        '401':
          description: Invalid/expired/reused refresh token (reuse revokes the family).
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
          x-error-codes: [INVALID_REFRESH_TOKEN, REFRESH_REUSE_DETECTED]
...
  /users/me/sessions:
    get:
      tags: [Users]
      operationId: listSessions
      summary: List active sessions.
      x-rate-limit: { policy: standard, limit: 60, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/SessionListEnvelope' } } }, description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/sessions/{sessionId}:
    delete:
      tags: [Users]
      operationId: revokeSession
      summary: Revoke a specific session.
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      x-rate-limit: { policy: standard, limit: 30, window: 1m }
      responses:
        '204': { description: Revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------- KYC -----
...
  /deposits/{depositId}:
    get:
      tags: [Deposits]
      operationId: getDeposit
      summary: Single deposit detail.
      parameters: [{ $ref: '#/components/parameters/DepositId' }]
      x-rate-limit: { policy: standard, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/DepositEnvelope' } } }, description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------- Withdrawals -----
...
  /inr/quotes:
    post:
      tags: [INR Ledger]
      operationId: createQuote
      summary: Get a short-lived signed price quote for INR↔USDT.
      x-rate-limit: { policy: trading, limit: 60, window: 1m }
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateQuoteRequest' } }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/QuoteEnvelope' } } }, description: Quote (expires shortly). }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /markets:
    get:
      tags: [Market Data]
      operationId: listMarkets
      summary: List tradable markets and their rules.
      security: []
      x-rate-limit: { policy: public, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/MarketListEnvelope' } } }, description: OK }
        '429': { $ref: '#/components/responses/RateLimited' }
  /markets/{symbol}/orderbook:
    get:
      tags: [Market Data]
      operationId: getOrderBook
      summary: Aggregated order-book depth (from the matching engine).
      security: []
      parameters:
        - { $ref: '#/components/parameters/MarketSymbol' }
        - name: depth
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      x-rate-limit: { policy: public, limit: 240, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/OrderBookEnvelope' } } }, description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
  /markets/{symbol}/ticker:
    get:
      tags: [Market Data]
      operationId: getTicker
      summary: 24h ticker for a market.
      security: []
      parameters: [{ $ref: '#/components/parameters/MarketSymbol' }]
      x-rate-limit: { policy: public, limit: 240, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/TickerEnvelope' } } }, description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
  /markets/{symbol}/candles:
    get:
      tags: [Market Data]
      operationId: getCandles
      summary: OHLCV candles.
      security: []
      parameters:
        - { $ref: '#/components/parameters/MarketSymbol' }
        - name: interval
          in: query
          required: true
          schema: { $ref: '#/components/schemas/CandleInterval' }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      x-rate-limit: { policy: public, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/CandleListEnvelope' } } }, description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List the user's orders.
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - name: market
          in: query
          schema: { type: string, example: BTC-USDT }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
      x-rate-limit: { policy: trading, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/OrderPageEnvelope' } } }, description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Orders]
      operationId: placeOrder
      summary: Place a spot order (locks funds, enters the matching engine).
      description: |
        LIMIT requires `price`+`quantity`. MARKET BUY requires `quoteBudget`;
        MARKET SELL requires `quantity`. STOP_LIMIT requires `stopPrice`+`price`+`quantity`.
        `clientOrderId` is an additional idempotency key per user.
      x-idempotent: required
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/PlaceOrderRequest' } }
      responses:
        '201': { content: { application/json: { schema: { $ref: '#/components/schemas/OrderEnvelope' } } }, description: Accepted/placed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Duplicate clientOrderId or idempotency in progress.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
          x-error-codes: [DUPLICATE_CLIENT_ORDER_ID, IDEMPOTENCY_IN_PROGRESS]
        '422':
          description: Validation, market rules, or balance failure.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
          x-error-codes: [VALIDATION_ERROR, INSUFFICIENT_BALANCE, MARKET_HALTED, TICK_SIZE_VIOLATION, STEP_SIZE_VIOLATION, MIN_NOTIONAL, POST_ONLY_WOULD_CROSS]
        '429': { $ref: '#/components/responses/RateLimited' }
  /orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrder
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      summary: Order detail.
      x-rate-limit: { policy: trading, limit: 240, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/OrderEnvelope' } } }, description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Orders]
      operationId: cancelOrder
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      summary: Cancel an open order (releases remaining lock).
      x-rate-limit: { policy: trading, limit: 120, window: 1m }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/OrderEnvelope' } } }, description: Cancelled. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Order is no longer cancellable.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
          x-error-codes: [ORDER_NOT_CANCELLABLE]

  # -------------------------------------------------------------- Trades -----
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access JWT (10–15 min TTL).
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access JWT on the admin host; TOTP mandatory at login.

  # ---- reusable parameters ----
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key per logical operation (8–255 chars). Replays return the original response.
      schema: { type: string, minLength: 8, maxLength: 255 }
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous `data.nextCursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    MarketSymbol:
      name: symbol
      in: path
      required: true
      schema: { type: string, pattern: '^[A-Z0-9]+-[A-Z0-9]+$', example: BTC-USDT }
    SessionId: { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
    DepositId: { name: depositId, in: path, required: true, schema: { type: string, format: uuid } }
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Unauthorized:
      description: Missing/invalid/expired token.
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: UNAUTHORIZED, message: Authentication required } } } }
...
    NotFound:
      description: Resource not found (or not visible to caller).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: NOT_FOUND, message: Resource not found } } } }
    ValidationError:
      description: Request failed schema validation (422). `details` lists field errors.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: VALIDATION_ERROR, message: Validation failed, details: [{ path: amount, message: must be > 0 }] } } } }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds until retry. }
        X-RateLimit-Remaining: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' }, example: { success: false, error: { code: RATE_LIMITED, message: Too many requests } } } }
...
  schemas:
...
    Decimal:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: Arbitrary-precision decimal as a string (never a float).
      example: '100.500000'
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
        details: { description: Optional structured context (e.g. field errors). }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: '#/components/schemas/Error' }
...
    DepositStatus: { type: string, enum: [DETECTED, CONFIRMING, CONFIRMED, CREDITED, ORPHANED] }
...
    OrderSide: { type: string, enum: [BUY, SELL] }
    OrderType: { type: string, enum: [LIMIT, MARKET, STOP_LIMIT] }
    TimeInForce: { type: string, enum: [GTC, IOC, FOK, POST_ONLY] }
    OrderStatus: { type: string, enum: [PENDING, OPEN, PARTIALLY_FILLED, FILLED, CANCELLED, REJECTED, EXPIRED] }
    MarketStatus: { type: string, enum: [ACTIVE, HALTED, POST_ONLY] }
    CandleInterval: { type: string, enum: [M1, M5, M15, H1, H4, D1, W1] }
    ConversionSide: { type: string, enum: [INR_TO_USDT, USDT_TO_INR] }
...
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 20 }
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { const: Bearer }
        expiresIn: { type: integer, description: Access token TTL (seconds). }
...
    Session:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        ip: { type: [string, 'null'] }
        device: { type: [object, 'null'] }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        current: { type: boolean }

    # ======================= DTOs: KYC =======================
...
    Deposit:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        chain: { type: string }
        asset: { type: string }
        amount: { $ref: '#/components/schemas/Decimal' }
        txHash: { type: string }
        confirmations: { type: integer }
        requiredConfirmations: { type: integer }
        status: { $ref: '#/components/schemas/DepositStatus' }
        detectedAt: { $ref: '#/components/schemas/Timestamp' }
        creditedAt: { type: [string, 'null'], format: date-time }

    # ======================= DTOs: Withdrawals =======================
...
    CreateQuoteRequest:
      type: object
      additionalProperties: false
      required: [side, amount]
      properties:
        side: { $ref: '#/components/schemas/ConversionSide' }
        amount: { allOf: [{ $ref: '#/components/schemas/Decimal' }], description: Source-asset amount (INR for INR_TO_USDT). }
    Quote:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        side: { $ref: '#/components/schemas/ConversionSide' }
        rate: { $ref: '#/components/schemas/Decimal' }
        spreadBps: { type: integer }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
...
    Market:
      type: object
      properties:
        symbol: { type: string }
        baseAsset: { type: string }
        quoteAsset: { type: string }
        status: { $ref: '#/components/schemas/MarketStatus' }
        tickSize: { $ref: '#/components/schemas/Decimal' }
        stepSize: { $ref: '#/components/schemas/Decimal' }
        minNotional: { $ref: '#/components/schemas/Decimal' }
        makerFeeBps: { type: integer }
        takerFeeBps: { type: integer }
    OrderBook:
      type: object
      properties:
        symbol: { type: string }
        sequence: { type: integer }
        bids: { type: array, items: { $ref: '#/components/schemas/PriceLevel' } }
        asks: { type: array, items: { $ref: '#/components/schemas/PriceLevel' } }
    PriceLevel:
      type: array
      description: '[price, quantity]'
      items: { $ref: '#/components/schemas/Decimal' }
      minItems: 2
      maxItems: 2
    Ticker:
      type: object
      properties:
        symbol: { type: string }
        lastPrice: { $ref: '#/components/schemas/Decimal' }
        bestBid: { $ref: '#/components/schemas/Decimal' }
        bestAsk: { $ref: '#/components/schemas/Decimal' }
        high24h: { $ref: '#/components/schemas/Decimal' }
        low24h: { $ref: '#/components/schemas/Decimal' }
        baseVolume24h: { $ref: '#/components/schemas/Decimal' }
        quoteVolume24h: { $ref: '#/components/schemas/Decimal' }
        priceChangePct: { $ref: '#/components/schemas/Decimal' }
    Candle:
      type: array
      description: '[openTime, open, high, low, close, baseVolume]'
      items: {}
      minItems: 6
      maxItems: 6
...
    PlaceOrderRequest:
      type: object
      additionalProperties: false
      required: [market, side, type]
      properties:
        market: { type: string, example: BTC-USDT }
        side: { $ref: '#/components/schemas/OrderSide' }
        type: { $ref: '#/components/schemas/OrderType' }
        tif: { $ref: '#/components/schemas/TimeInForce' }
        price: { $ref: '#/components/schemas/Decimal' }
        stopPrice: { $ref: '#/components/schemas/Decimal' }
        quantity: { $ref: '#/components/schemas/Decimal' }
        quoteBudget: { allOf: [{ $ref: '#/components/schemas/Decimal' }], description: For MARKET BUY — quote amount to spend. }
        clientOrderId: { type: string, maxLength: 64, description: Per-user idempotency token. }
      allOf:
        - description: |
            Conditional requirements (enforced server-side):
            LIMIT → price & quantity; MARKET BUY → quoteBudget; MARKET SELL → quantity;
            STOP_LIMIT → stopPrice & price & quantity. Price/qty must satisfy the
            market's tickSize/stepSize/minNotional.
    Order:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        clientOrderId: { type: [string, 'null'] }
        market: { type: string }
        side: { $ref: '#/components/schemas/OrderSide' }
        type: { $ref: '#/components/schemas/OrderType' }
        tif: { $ref: '#/components/schemas/TimeInForce' }
        price: { type: [string, 'null'] }
        stopPrice: { type: [string, 'null'] }
        quantity: { type: [string, 'null'] }
        filledQuantity: { $ref: '#/components/schemas/Decimal' }
        quoteSpent: { $ref: '#/components/schemas/Decimal' }
        status: { $ref: '#/components/schemas/OrderStatus' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        closedAt: { type: [string, 'null'], format: date-time }
...
    TokenPairEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/TokenPair' } } }
...
    SessionListEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Session' } } } } } }
...
    DepositEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/Deposit' } } }
...
    QuoteEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/Quote' } } }
...
    MarketListEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Market' } } } } } }
    OrderBookEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/OrderBook' } } }
    TickerEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/Ticker' } } }
    CandleListEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { type: object, properties: { interval: { $ref: '#/components/schemas/CandleInterval' }, items: { type: array, items: { $ref: '#/components/schemas/Candle' } } } } } }
...
    OrderEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { $ref: '#/components/schemas/Order' } } }
    OrderPageEnvelope: { type: object, required: [success, data], properties: { success: { const: true }, data: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Order' } }, nextCursor: { type: [string, 'null'] } } } } }
...